% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprLogLik.R
\name{exprLogLik}
\alias{exprLogLik}
\title{Calculate Log-Likelihood-Ratio for Expression data}
\usage{
exprLogLik(expr, param, same_sd = T, k_win = 2.5)
}
\description{
To ensure robustness the log-lik is linearized when
the expression measurement is more than k_win standard
deviations away from the group mean.
}
