% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{exprPrior}
\alias{exprPrior}
\title{Find prior for expression data}
\usage{
exprPrior(exprMethList, bins = 10, plot = TRUE)
}
\arguments{
\item{exprMethList}{List of nx3 matrices with expression measurements in the first column and
methylation measurements in second and third column}
}
\description{
Find prior for expression data
}
