% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitExpr.R
\name{fitExpr}
\alias{fitExpr}
\title{Fit expression data}
\usage{
fitExpr(expr, grouping, prior = NULL, k_win = 2.5, pooled = T)
}
\arguments{
\item{expr}{vector with expression values}

\item{grouping}{logical vector indicating group relationship}

\item{k_win}{amount of winsorization cap values more than k_win standard deviations away from mean}

\item{pooled}{use a common standard deviation estimate for each group}
}
\description{
Fit expression data
}
