% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMeth.R
\name{fitMeth}
\alias{fitMeth}
\title{Fit expression data Linear Normal Model}
\usage{
fitMeth(meth, expr, grouping, win = F, prior = NULL)
}
\arguments{
\item{meth}{matrix with met}

\item{expr}{vector with expression values}

\item{grouping}{logical vector indicating group relationship}

\item{meth}{matrix with methylation values}
}
\description{
Fit expression data Linear Normal Model
}
