% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl.R
\name{kl_multivariate}
\alias{kl_multivariate}
\title{KL Multivariate Normal
Compute KL-divergence between two
multivariate normal distributions}
\usage{
kl_multivariate(mu1, Sigma1, mu2, Sigma2)
}
\description{
KL Multivariate Normal
Compute KL-divergence between two
multivariate normal distributions
}
