% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl.R
\name{kl_normal}
\alias{kl_normal}
\title{KL Univariate Normal
Compute KL-divergence between two
univariate normal distributions}
\usage{
kl_normal(mu1, sigma1, mu2, sigma2)
}
\description{
KL Univariate Normal
Compute KL-divergence between two
univariate normal distributions
}
