% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methLogLik.R
\name{methLogLik}
\alias{methLogLik}
\title{Calculate Log-Likelihood-Ratio for Expression data}
\usage{
methLogLik(meth, expr, param, same_cov = T)
}
\description{
Calculate Log-Likelihood-Ratio for Expression data
}
