% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{methPrior}
\alias{methPrior}
\title{Find prior for methylation data}
\usage{
methPrior(exprMethList)
}
\arguments{
\item{exprMethList}{List of nx3 matrices with expression measurements in the first column and
methylation measurements in second and third column}
}
\description{
Find prior for methylation data
}
