% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDE.R
\name{testDE}
\alias{testDE}
\title{Test for differential Expression}
\usage{
testDE(expr, grouping, prior = NULL, k_win = 2.5, pooled = F,
  winz_df = T)
}
\arguments{
\item{expr}{Vector with expression values}

\item{grouping}{Logical vector indicating group relationship}

\item{prior}{}
}
\description{
Fit expression model and evaluate differential expression
Robust regression model with prior on standard deviation
}
