% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsorize.R
\name{winsorize}
\alias{winsorize}
\title{Winsorize Grouped Observations}
\usage{
winsorize(x, g, k_win, max_iter = 10, eps = 1e-04, boost = FALSE,
  pooled = TRUE)
}
\arguments{
\item{x}{observations to be winsorized}

\item{g}{logical vector indicating group relationship}

\item{k_win}{degree of winsorization}

\item{eps}{convergence tolerance in winsorization procedure}

\item{boost}{multiply difference from mean by a factor to reduce bias on scale estimate(s)}

\item{pooled}{use a common standard deviation estimate for each group}
}
\description{
Winsorize Grouped Observations
}
